Historical Fallout 13 capitulation
On this day May 23 - 2017, in order to let players continue to enjoy the Fallout 13 project and allow the further hosting of our servers on Byond HUB, Fallout 13 administration has made an uneasy decision to ease monopoly on the source code.
Developers have tried their best to defend the project, by seeking holes in licensing, and other ways not to surrender, but capitulation turned out the only reasonable thing to do, besides a complete termination of a project.
This day shall be remembered as a day of great shame and disappointment, but also a day when each and everyone got the ability to directly contribute to the project, a day when Fallout 13 was saved from a Great War of Communities, and potential extinction by the course of war.